+++ src/migrations/index.ts
@@ -2,6 +2,7 @@ import * as migration_20260906_040011 from './20260906_040011';
 import * as migration_20260911_113500_add_homepage_font_family from './20260911_113500_add_homepage_font_family';
 import * as migration_20260912_102452 from './20260912_102452';
 import * as migration_20260912_104426_scientific_activity_groups from './20260912_104426_scientific_activity_groups';
+import * as migration_20260913_124616 from './20260913_124616';
 
 export const migrations = [
   {
@@ -23,5 +24,10 @@ export const migrations = [
     up: migration_20260912_104426_scientific_activity_groups.up,
     down: migration_20260912_104426_scientific_activity_groups.down,
     name: '20260912_104426_scientific_activity_groups',
+  },
+  {
+    up: migration_20260913_124616.up,
+    down: migration_20260913_124616.down,
+    name: '20260913_124616'
   },
 ];